+++ CotEditor/Sources/DefaultKey.swift
@@ -270,8 +270,6 @@ extension DefaultKeys {
     static let headerFooterDateFormat = DefaultKey<String>("headerFooterDateFormat")
     static let headerFooterPathAbbreviatingWithTilde = DefaultKey<Bool>("headerFooterPathAbbreviatingWithTilde")
     static let autoCompletionDelay = DefaultKey<Double>("autoCompletionDelay")
-    static let infoUpdateInterval = DefaultKey<Double>("infoUpdateInterval")
-    static let outlineMenuInterval = DefaultKey<Double>("outlineMenuInterval")
     static let showColoringIndicatorTextLength = DefaultKey<Int>("showColoringIndicatorTextLength")
     static let coloringRangeBufferLength = DefaultKey<Int>("coloringRangeBufferLength")
     static let largeFileAlertThreshold = DefaultKey<Int>("largeFileAlertThreshold")